A running version of new JobManager
