+++ ApsimNG/Commands/RunCommand.cs
@@ -55,7 +55,7 @@
             jobManager = Runner.ForSimulations(explorerPresenter.ApsimXFile, model, false);
 
             if (multiProcess)
-                jobRunner = new JobManagerMultiProcess(storage);
+                jobRunner = new JobRunnerMultiProcess(storage);
             else
                 jobRunner = new JobRunnerAsync();
             jobRunner.JobCompleted += OnJobCompleded;
@@ -98,7 +98,7 @@
                 errors.Add(e.exceptionThrown);
 
             Stop();
-            if (errors.Count > 0)
+            if (errors.Count == 0)
                 explorerPresenter.MainPresenter.ShowMessage(jobName + " complete "
                         + " [" + stopwatch.Elapsed.TotalSeconds.ToString("#.00") + " sec]", Simulation.ErrorLevel.Information);
             else
@@ -107,7 +107,6 @@
                 errors.ForEach(error => errorMessage += error.ToString() + Environment.NewLine);
                 explorerPresenter.MainPresenter.ShowMessage(errorMessage, Simulation.ErrorLevel.Error);
             }
-            explorerPresenter.MainPresenter.ShowProgress(100);
 
             SoundPlayer player = new SoundPlayer();
             if (DateTime.Now.Month == 12 && DateTime.Now.Day == 25)
